# brokerage-apis/openapi/brokerage-api.yaml
openapi: 3.2.1
...
info:
  title: Brokerage API
  version: "2.0"
  summary: Accounts, positions, market data and trading for a retail brokerage, as hypermedia.
  description: |
    A resource-oriented REST API that follows the standards in *REST API Standards with
    HATEOAS* (Smiroh Publishers).

    * **Hypermedia.** Every representation is HAL (`application/hal+json`) with a `_links`
      object. Ask for `application/prs.hal-forms+json` to also receive `_templates`: the
      state transitions available right now. Clients should bookmark only the API root and
      follow links from there. Custom link relations use the CURIE `bk:`, which expands to
      `https://docs.brokerage.example/rels/{rel}`.
    * **Money.** Prices, amounts and quantities are decimal strings, never JSON numbers.
    * **Errors.** Every error is an RFC 9457 problem document (`application/problem+json`)
      whose `type` URI identifies the error. Branch on `type`, never on `title` or `detail`.
    * **Versions.** Send `API-Version: 2` to opt in to version 2. Version 1 is the default and
      is deprecated: its responses carry `Deprecation` (RFC 9745) and `Sunset` (RFC 8594).
    * **Quotas.** Every response carries `RateLimit-Policy` and `RateLimit`. Exceeding a quota
      returns 429 with `Retry-After`.
    * **Tracing.** Every response echoes `X-Request-Id`; quote it in support requests.
  contact:
    name: Brokerage API support
    url: https://docs.brokerage.example/support
  license:
...
servers:
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      summary: List accounts
      tags: [accounts]
      security: [{ oauth2: [accounts:read] }]
      responses:
        "200":
          description: The caller's accounts. Links depend on the token's scopes.
          headers:
            RateLimit: { $ref: "#/components/headers/RateLimit" }
            RateLimit-Policy: { $ref: "#/components/headers/RateLimit-Policy" }
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/AccountCollection" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /accounts/{accountId}:
    parameters:
      - $ref: "#/components/parameters/accountId"
    get:
      operationId: getAccount
      summary: Get an account
      tags: [accounts]
      security: [{ oauth2: [accounts:read] }]
      responses:
        "200":
          description: >-
            The account. With orders:read the links include orders, previews and events; with
            orders:write, HAL-FORMS clients also receive a placeOrder template.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Account" }
            application/prs.hal-forms+json:
              schema: { $ref: "#/components/schemas/Account" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /accounts/{accountId}/balances:
    parameters:
      - $ref: "#/components/parameters/accountId"
    get:
      operationId: getBalances
      summary: Get balances and buying power
      tags: [holdings]
      security: [{ oauth2: [accounts:read] }]
      responses:
        "200":
          description: Cash, reservations for open orders, buying power and market value.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Balances" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /accounts/{accountId}/positions:
    parameters:
      - $ref: "#/components/parameters/accountId"
    get:
      operationId: listPositions
      summary: List positions
      tags: [holdings]
      security: [{ oauth2: [accounts:read] }]
      responses:
        "200":
          description: Every holding, valued at the current market price.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/PositionCollection" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /accounts/{accountId}/positions/{instrumentId}:
    parameters:
      - $ref: "#/components/parameters/accountId"
      - $ref: "#/components/parameters/instrumentId"
    get:
      operationId: getPosition
      summary: Get one position
      tags: [holdings]
      security: [{ oauth2: [accounts:read] }]
      responses:
        "200":
          description: The position in one instrument.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Position" }
        "404": { $ref: "#/components/responses/NotFound" }
  /accounts/{accountId}/transactions:
    parameters:
      - $ref: "#/components/parameters/accountId"
    get:
      operationId: listTransactions
      summary: List transactions
      description: Newest first. Follow the `next` link for older entries.
      tags: [transactions]
      security: [{ oauth2: [accounts:read] }]
      parameters:
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/TransactionType" }
        - name: from
          in: query
          description: Earliest date to include (UTC), inclusive.
          schema: { type: string, format: date }
        - name: to
          in: query
          description: Latest date to include (UTC), inclusive.
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: One page of transactions.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/TransactionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /accounts/{accountId}/orders:
    parameters:
      - $ref: "#/components/parameters/accountId"
    get:
      operationId: listOrders
      summary: List orders
      description: Newest first, optionally filtered. `status` may be repeated.
      tags: [orders]
      security: [{ oauth2: [orders:read] }]
      parameters:
        - name: status
          in: query
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: "#/components/schemas/OrderStatus" }
        - name: symbol
          in: query
          schema: { type: string, examples: [AAPL] }
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: One page of orders.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/OrderPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: placeOrder
      summary: Place an order
      description: >-
        Every new order needs a unique Idempotency-Key. Retrying with the same key and body
        returns the original order (Idempotent-Replayed: true); the same key with a different
        body is refused.
      tags: [orders]
      security: [{ oauth2: [orders:write] }]
      parameters:
        - $ref: "#/components/parameters/Idempotency-Key"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderRequest" }
            examples:
              limitBuy:
                summary: Buy 10 AAPL at 230.00 or better
                dataValue:
                  instrumentId: EQ-AAPL
                  side: BUY
                  type: LIMIT
                  timeInForce: DAY
                  quantity: "10"
                  limitPrice: "230.00"
                  clientOrderId: rebalance-2026-09
      responses:
        "201":
          description: The order was accepted and is now working.
          headers:
            Location: { $ref: "#/components/headers/Location" }
            ETag: { $ref: "#/components/headers/ETag" }
            Idempotent-Replayed: { $ref: "#/components/headers/Idempotent-Replayed" }
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Order" }
          links:
            getOrder:
              operationId: getOrder
              parameters:
                accountId: $request.path.accountId
                orderId: $response.body#/id
            cancelOrder:
              operationId: cancelOrder
              parameters:
                accountId: $request.path.accountId
                orderId: $response.body#/id
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableContent" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /accounts/{accountId}/order-previews:
    parameters:
      - $ref: "#/components/parameters/accountId"
    post:
      operationId: previewOrder
      summary: Preview an order
      description: >-
        Estimates cost and checks every rule without placing anything. A preview that would be
        refused is still a 200, with `acceptable: false` and the reasons.
      tags: [orders]
      security: [{ oauth2: [orders:read] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderRequest" }
      responses:
        "200":
          description: The estimate.
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/OrderPreview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: >-
        OAuth 2.1. Web apps sign customers in through their backend, a confidential client,
        with the authorization code grant and PKCE; a customer's own scripts and agents use
        client credentials registered to that customer. Tokens carry `customer_id` and, for
        delegated clients, `trading_limit`.
      oauth2MetadataUrl: http://localhost:9000/.well-known/oauth-authorization-server
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:9000/oauth2/authorize
          tokenUrl: http://localhost:9000/oauth2/token
          refreshUrl: http://localhost:9000/oauth2/token
          scopes:
            market-data:read: Instruments, quotes and option chains
            accounts:read: Accounts, balances, positions and transactions
            orders:read: Orders, previews and order events
            orders:write: Place, change and cancel orders
        clientCredentials:
          tokenUrl: http://localhost:9000/oauth2/token
          scopes:
            market-data:read: Instruments, quotes and option chains
            accounts:read: Accounts, balances, positions and transactions
            orders:read: Orders, previews and order events
            orders:write: Place, change and cancel orders
  parameters:
    accountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, examples: [ACC-1001] }
...
    instrumentId:
      name: instrumentId
      in: path
      required: true
      schema: { type: string, examples: [EQ-AAPL] }
    cursor:
      name: cursor
      in: query
      description: Opaque. Take it from a `next` link; never build one.
      schema: { type: string }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Idempotency-Key:
      name: Idempotency-Key
      in: header
      required: true
      description: A unique value per new order (a UUID works well). Retries reuse it.
      schema: { type: string, minLength: 1, maxLength: 255 }
...
  headers:
...
    Location:
      description: The URL of the resource created or affected.
      schema: { type: string, format: uri }
    ETag:
      description: The representation's entity tag.
      schema: { type: string }
    Idempotent-Replayed:
      description: true when the response replays an earlier request with the same Idempotency-Key.
      schema: { type: string, enum: ["true", "false"] }
    RateLimit:
      description: 'Remaining quota and seconds until reset, e.g. "default";r=117;t=42.'
      schema: { type: string }
    RateLimit-Policy:
      description: 'The quota policy, e.g. "default";q=120;w=60.'
      schema: { type: string }
    Retry-After:
      description: Seconds to wait before retrying.
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            invalidFields:
              dataValue:
                type: https://docs.brokerage.example/problems/invalid-request
                title: The request is not valid
                status: 400
                detail: One or more fields are invalid
                instance: /accounts/ACC-1001/orders
                requestId: 7f3c2a90-4a1e-4c1b-9d3e-2b8f1c6e0a11
                errors:
                  - { field: quantity, message: must be greater than 0 }
    Unauthorized:
      description: No valid bearer token. See WWW-Authenticate.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The token lacks the scope, or the order exceeds the client's trading limit.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: The resource does not exist, or belongs to another customer.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    UnprocessableContent:
      description: The request is well formed but breaks a business rule.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            buyingPower:
              dataValue:
                type: https://docs.brokerage.example/problems/insufficient-buying-power
                title: The account does not have enough buying power
                status: 422
                detail: This order needs 250000.00 but only 52913.10 is available
                instance: /accounts/ACC-1001/orders
                requestId: b1b72c27-9103-4236-9cc0-c2b21d4d86bf
                required: "250000.00"
                available: "52913.10"
    TooManyRequests:
      description: The client's quota is used up.
      headers:
        Retry-After: { $ref: "#/components/headers/Retry-After" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Decimal:
      description: An exact decimal number, as a string.
      type: string
      pattern: '^-?[0-9]+(\.[0-9]+)?$'
      examples: ["187.50"]
    Link:
      type: object
      required: [href]
      properties:
        href: { type: string }
        templated: { type: boolean }
        type: { type: string }
        name: { type: string }
        title: { type: string }
    Links:
      description: HAL links, keyed by relation. A relation may hold one link or several.
      type: object
      properties:
        self: { $ref: "#/components/schemas/Link" }
        curies:
          type: array
          items: { $ref: "#/components/schemas/Link" }
      additionalProperties:
        anyOf:
          - $ref: "#/components/schemas/Link"
          - type: array
            items: { $ref: "#/components/schemas/Link" }
    Templates:
      description: HAL-FORMS templates, present only for application/prs.hal-forms+json.
      type: object
      additionalProperties:
        type: object
        required: [method]
        properties:
          method: { type: string }
          contentType: { type: string }
          target: { type: string }
          properties: { type: array, items: { type: object } }
    Problem:
      description: An RFC 9457 problem document. Extension members depend on the type.
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    Account:
      type: object
      required: [id, number, type, registration, currency, _links]
      properties:
        id: { type: string }
        number:
          type: string
          description: Masked; only the last four digits are sent
          examples: ["****4821"]
        nickname: { type: string }
        type: { type: string, enum: [CASH, MARGIN] }
        registration: { type: string, enum: [INDIVIDUAL, JOINT, ROTH_IRA, TRADITIONAL_IRA] }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        openedOn: { type: string, format: date }
        _links: { $ref: "#/components/schemas/Links" }
        _templates: { $ref: "#/components/schemas/Templates" }
    AccountCollection:
      type: object
      required: [_links]
      properties:
        _embedded:
          type: object
          properties:
            bk:accounts: { type: array, items: { $ref: "#/components/schemas/Account" } }
        _links: { $ref: "#/components/schemas/Links" }
    Balances:
      type: object
      required: [accountId, currency, cash, reservedForOrders, buyingPower, positionsMarketValue,
        totalEquity, unrealizedPnl, asOf, _links]
      properties:
        accountId: { type: string }
        currency: { type: string }
        cash: { $ref: "#/components/schemas/Decimal" }
        reservedForOrders: { $ref: "#/components/schemas/Decimal" }
        buyingPower: { $ref: "#/components/schemas/Decimal" }
        positionsMarketValue: { $ref: "#/components/schemas/Decimal" }
        totalEquity: { $ref: "#/components/schemas/Decimal" }
        unrealizedPnl: { $ref: "#/components/schemas/Decimal" }
        asOf: { type: string, format: date-time }
        _links: { $ref: "#/components/schemas/Links" }
    Position:
      type: object
      required: [instrumentId, symbol, instrumentType, quantity, averageCost, costBasis, lastPrice,
        marketValue, unrealizedPnl, unrealizedPnlPercent, _links]
      properties:
        instrumentId: { type: string }
        symbol: { type: string }
        instrumentType: { $ref: "#/components/schemas/InstrumentType" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        averageCost: { $ref: "#/components/schemas/Decimal" }
        costBasis: { $ref: "#/components/schemas/Decimal" }
        lastPrice: { $ref: "#/components/schemas/Decimal" }
        marketValue: { $ref: "#/components/schemas/Decimal" }
        unrealizedPnl: { $ref: "#/components/schemas/Decimal" }
        unrealizedPnlPercent: { $ref: "#/components/schemas/Decimal" }
        _links: { $ref: "#/components/schemas/Links" }
    PositionCollection:
      type: object
      required: [_links]
      properties:
        _embedded:
          type: object
          properties:
            bk:positions: { type: array, items: { $ref: "#/components/schemas/Position" } }
        _links: { $ref: "#/components/schemas/Links" }
    TransactionType:
      type: string
      enum: [DEPOSIT, WITHDRAWAL, TRADE, DIVIDEND, FEE]
    Transaction:
      type: object
      required: [id, type, description, amount, occurredAt, settlesOn]
      properties:
        id: { type: string }
        type: { $ref: "#/components/schemas/TransactionType" }
        description: { type: string }
        amount: { $ref: "#/components/schemas/Decimal", description: Positive into the account, negative out }
        instrumentId: { type: string }
        symbol: { type: string }
        quantity: { $ref: "#/components/schemas/Decimal" }
        price: { $ref: "#/components/schemas/Decimal" }
        orderId: { type: string }
        occurredAt: { type: string, format: date-time }
        settlesOn: { type: string, format: date }
    TransactionPage:
      type: object
      required: [_links]
      properties:
        _embedded:
          type: object
          properties:
            bk:transactions: { type: array, items: { $ref: "#/components/schemas/Transaction" } }
        _links: { $ref: "#/components/schemas/Links" }

    # ------------------------------------------------------------------ market data
    InstrumentType:
      type: string
      enum: [EQUITY, ETF, OPTION]
...
    OrderStatus:
      type: string
      enum: [OPEN, PARTIALLY_FILLED, FILLED, PENDING_CANCEL, CANCELLED, REJECTED]
    OrderSide:
      type: string
      enum: [BUY, SELL]
    OrderType:
      type: string
      enum: [MARKET, LIMIT, STOP, STOP_LIMIT]
    TimeInForce:
      type: string
      enum: [DAY, GTC]
      default: DAY
    OrderRequest:
      type: object
      required: [instrumentId, side, type, quantity]
      properties:
        instrumentId: { type: string, minLength: 1 }
        side: { $ref: "#/components/schemas/OrderSide" }
        type: { $ref: "#/components/schemas/OrderType" }
        timeInForce: { $ref: "#/components/schemas/TimeInForce" }
        quantity: { $ref: "#/components/schemas/Decimal", description: Shares, or contracts for options }
        limitPrice: { $ref: "#/components/schemas/Decimal", description: Required for LIMIT and STOP_LIMIT }
        stopPrice: { $ref: "#/components/schemas/Decimal", description: Required for STOP and STOP_LIMIT }
        clientOrderId: { type: string, maxLength: 64, pattern: "^[A-Za-z0-9._-]*$" }
...
    Fill:
      type: object
      required: [id, quantity, price, executedAt]
      properties:
        id: { type: string }
        quantity: { $ref: "#/components/schemas/Decimal" }
        price: { $ref: "#/components/schemas/Decimal" }
        executedAt: { type: string, format: date-time }
    Order:
      type: object
      required: [id, accountId, instrumentId, symbol, instrumentType, side, type, timeInForce, quantity,
        status, filledQuantity, remainingQuantity, fills, createdAt, updatedAt, _links]
      properties:
        id: { type: string }
        accountId: { type: string }
        clientOrderId: { type: string }
        instrumentId: { type: string }
        symbol: { type: string }
        instrumentType: { $ref: "#/components/schemas/InstrumentType" }
        side: { $ref: "#/components/schemas/OrderSide" }
        type: { $ref: "#/components/schemas/OrderType" }
        timeInForce: { $ref: "#/components/schemas/TimeInForce" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        limitPrice: { $ref: "#/components/schemas/Decimal" }
        stopPrice: { $ref: "#/components/schemas/Decimal" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        filledQuantity: { $ref: "#/components/schemas/Decimal" }
        remainingQuantity: { $ref: "#/components/schemas/Decimal" }
        averageFillPrice: { $ref: "#/components/schemas/Decimal" }
        fills: { type: array, items: { $ref: "#/components/schemas/Fill" } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        _links: { $ref: "#/components/schemas/Links" }
        _templates: { $ref: "#/components/schemas/Templates" }
    OrderPage:
      type: object
      required: [_links]
      properties:
        _embedded:
          type: object
          properties:
            bk:orders: { type: array, items: { $ref: "#/components/schemas/Order" } }
        _links: { $ref: "#/components/schemas/Links" }
    OrderPreview:
      type: object
      required: [accountId, instrumentId, side, type, quantity, acceptable, issues]
      properties:
        accountId: { type: string }
        instrumentId: { type: string }
        symbol: { type: string }
        side: { $ref: "#/components/schemas/OrderSide" }
        type: { $ref: "#/components/schemas/OrderType" }
        quantity: { $ref: "#/components/schemas/Decimal" }
        estimatedPrice: { $ref: "#/components/schemas/Decimal" }
        estimatedNotional: { $ref: "#/components/schemas/Decimal" }
        estimatedCommission: { $ref: "#/components/schemas/Decimal" }
        estimatedTotal: { $ref: "#/components/schemas/Decimal" }
        buyingPower: { $ref: "#/components/schemas/Decimal" }
        acceptable: { type: boolean }
        issues:
          type: array
          items:
            type: object
            required: [type, detail]
            properties:
              type: { type: string, format: uri }
              detail: { type: string }
        _links: { $ref: "#/components/schemas/Links" }
...